requirements.txt
PySide6>=6.6.0
psycopg2-binary>=2.9.9
python-barcode>=0.15.1
reportlab>=4.0.7
Pillow>=10.1.0
python-dotenv>=1.0.0
bcrypt>=4.1.1
pandas>=2.1.3
openpyxl>=3.1.2
qrcode>=7.4.2
